backend/requirements.txt
flask==3.0.0
flask-cors
pandas
numpy
scikit-learn==1.3.2
openpyxl
google-generativeai>=0.5.0
python-dotenv
torch
pytorch-tabnet==3.1.1
hdbscan
joblib
loguru
scipy
